requirements.txt
# SecureSight Backend Dependencies

# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyotp>=2.9.0
qrcode[pil]>=7.4.2

# Database
sqlalchemy>=2.0.25
asyncpg>=0.30.0
alembic>=1.13.1
psycopg[binary]>=3.1.0

# Elasticsearch
elasticsearch[async]>=8.12.0,<9.0.0

# Redis/Message Queue
redis>=5.0.1
aioredis>=2.0.1

# Kafka (optional)
aiokafka>=0.11.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.11.0

# Email
aiosmtplib>=3.0.1

# Validation & Parsing
email-validator>=2.1.0
python-dateutil>=2.8.2

# Logging & Monitoring
structlog>=24.1.0
prometheus-client>=0.19.0

# Rate Limiting
slowapi>=0.1.9

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Development
black>=24.1.0
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
orjson>=3.10.0
